Fix new message do not appear sometimes (#185)
